add "Knowledge" story type for intangible matchers
Useful for seeing if the player has learned about something, or if an event has happened, or even giving the player new skills.

The size of this commit shows the pain in adding a new story type with the current architechture and type design.
That seems unfortunate, although I am modeling a new concept in the domain, so maybe it is necessary?